% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize.R
\name{categorize}
\alias{categorize}
\title{categorize}
\usage{
categorize(data, vars, method = c("reference", "cellmeans"), ref = c())
}
\arguments{
\item{data}{input data frame with no NAs, should include all the covariates that will be used later}

\item{vars}{name vector for the category variables}

\item{method}{specify the method for categorization}

\item{refs}{specify the reference level for each categorical variable}
}
\value{
data frame with all the category variables transformed to numerous dummy variables.
}
\description{
categorize a set of variables automatically into dummy variables. Return data frame with categorized
values. The method could only be reference coding or cell means coding.Score coding is not integrated
in this function. The main usage for this function is usesd internally in mylm function. But user could
also use this to create categorical variables as well.
}
\examples{
data(mydata)
library(qwzLinear)
categorize(mydata, c('R_E', 'NIHSS_4Cat'), ref = c(0,1))
categorize(mydata, c('R_E', 'NIHSS_4Cat'), method = 'cellmeans') = F)

}
