% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{Linear Regression Model}
\usage{
linear_regression(x, y)
}
\arguments{
\item{x}{A numeric vector of predictor values.}

\item{y}{A numeric vector of response values.}
}
\value{
A list with the following components:
\item{intercept }{The estimated value of the intercept in the linear regression model.}
\item{slope }{The estimated value of the slope in the linear regression model.}
}
\description{
This function fits a simple linear regression model by estimating the coefficients for the intercept and slope of a line that best fits the input data.
}
\details{
The function estimates the linear relationship between the predictor \code{x} and response \code{y} by calculating the slope and intercept. The slope is calculated as the covariance of \code{x} and \code{y} divided by the variance of \code{x}. The intercept is calculated as the difference between the mean of \code{y} and the product of the slope with the mean of \code{x}.
}
\examples{
# Example usage of linear_regression
x <- 1:10
y <- 2 * x + rnorm(10)
model <- linear_regression(x, y)
print(model)

}
