% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{The Stroke dataset used in BIOSTATS-650}
\format{
A data frame with 612 rows and 53 columns
}
\source{
Morgenstern, Lewis B., et al. "Fatalism, optimism, spirituality, depressive symptoms, and stroke
outcome: a population-based analysis." Stroke 42.12 (2011): 3518-3523.
}
\usage{
mydata
}
\description{
The Stroke dataset used in BIOSTATS-650
}
\keyword{datasets}
